+++ includes/functions-display.php
@@ -3,6 +3,9 @@
  * @package User_Locations
  */
 
+// Exit if accessed directly
+if ( ! defined( 'ABSPATH' ) ) exit;
+
 // Create shortcode functionality. Functions are defined in includes/wpseo-local-functions.php because they're also used by some widgets.
 add_shortcode( 'ul_address',       'ul_show_address' );
 add_shortcode( 'ul_map',           'ul_show_map' );